feat: add connect to supabase

+++ modules/entities/user.go
@@ -1,7 +1,7 @@
 package entities
 
 type UsersRegisterRes struct {
-	Id       uint64 `json:"id" db:"id"`
+	Id       int8   `json:"id" db:"id"`
 	Username string `json:"username" db:"username"`
 }
